Delete commented lines and fix event date parsing and storage format
