+++ data/koji.txt
@@ -1,3 +1,3 @@
 D | 0 | return book | 2022-08-10 1800
 T | 0 | abcdefghijksj
-E | 0 | study sesh | Feb 01 2025, 9:00 am | Mar 01 2025, 7:00 pm
+E | 0 | study sesh | 2025-02-01 0900 | 2025-03-01 1900
