--- src/main/java/koji/KojiException.java
@@ -1,7 +1,0 @@
-package koji;
-
-public class KojiException extends Exception {
-    public KojiException(String message) {
-        super(message);
-    }
-}
