+++ src/main/java/task/Event.java
@@ -28,6 +28,6 @@ public class Event extends Task {
 
     @Override
     public String toFileString() {
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from.format(OUTPUT_FORMAT) + " | " + to.format(OUTPUT_FORMAT);
+        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from.format(INPUT_FORMAT) + " | " + to.format(INPUT_FORMAT);
     }
 }
